refactor(context): move context to pkg
