+++ internal/api/handlers_objects.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"time"
 
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
 	"nimona.io/internal/http/router"
 	"nimona.io/internal/store/kv"
