+++ internal/log/writer_json.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 )
 
 func JSONWriter() Writer {
