+++ pkg/discovery/hyperspace/resolver.go
@@ -5,7 +5,7 @@ import (
 
 	"nimona.io/pkg/object"
 
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
 	"nimona.io/internal/log"
 	"nimona.io/pkg/crypto"
