+++ pkg/discovery/hyperspace/resolver_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 	"nimona.io/internal/store/graph"
 	"nimona.io/internal/store/kv"
 	"nimona.io/pkg/crypto"
