+++ pkg/exchange/exchange.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/gobwas/glob"
 
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
 	"nimona.io/internal/log"
 	"nimona.io/internal/store/graph"
