+++ pkg/orchestrator/orchestrator_sync.go
@@ -1,7 +1,7 @@
 package orchestrator
 
 import (
-	"nimona.io/internal/context"
+	"nimona.io/pkg/context"
 	"nimona.io/internal/errors"
 	"nimona.io/internal/log"
 	"nimona.io/pkg/exchange"
